pyproject.toml
[tool.poetry]
authors = ["bruvio <bruno.viola@pm.me>"]
description = "Django website for a personal portfolio"
license = "MIT"
name = "personal_portfolio-bruvio"
version = "1.1.0"

[tool.poetry.dependencies]
Django = "3.2.8"
Pillow = "^9.3.0"
boto3 = "^1.26.7"
django-filter = "^21.1"
django-storages = "^1.13.1"
djangorestframework = "^3.13.1"
psycopg2-binary = "^2.9.3"
python = "^3.8"
python-decouple = "^3.6"
uWSGI = "^2.0.20"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
flake8 = "^4.0.1"
isort = "^5.10.1"
pep8-naming = "^0.12.1"
pre-commit = "^2.19.0"
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
pytest-django = "^4.5.2"
pytest-flake8 = "^1.1.1"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.isort]
ensure_newline_before_comments = true
force_grid_wrap = 0
include_trailing_comma = true
line_length = 79
multi_line_output = 3
profile = 'black'
use_parentheses = true

[tool.black]
line-length = 79

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "personal_portfolio.settings"
addopts = '-s -v'
